requirements.txt
dash
flask
werkzeug
click>=8.0
markupsafe
python-dotenv
requests
